requirements.txt
scikit-learn==0.22.dev0
numpy==1.18.1
scipy==1.4.1
pot==0.5.1
pytorch==1.5.0
netcdf4==1.4.2
torchvision==0.6.0
pillow==6.1.0
mpl_toolkits
seaborn==0.10.1
pandas==1.0.3
matplotlib==3.1.3
mpl_toolkits==1.2.0
